Change: domain, infra層のディレクトリ構成変更
