+++ __test__/unit/application/usecases/read-all-users-usecase-impl.test.ts
@@ -3,7 +3,7 @@ import { ReadAllUsersUseCaseImpl } from "../../../../src/application/usecases/re
 import { InternalError } from "../../../../src/core/error/internal-error";
 import { Failure, Success } from "../../../../src/core/result";
 import { User } from "../../../../src/domain/entities/user";
-import { UserRepository } from "../../../../src/domain/repositories/user-repository";
+import { UserRepository } from "../../../../src/domain/interfaces/repositories/user-repository";
 
 describe("ReadAllUsersUseCase", () => {
   const users: User[] = [
