+++ src/application/usecases/delete-user-usecase-impl.ts
@@ -2,11 +2,11 @@ import { inject, injectable } from "inversify";
 import "reflect-metadata";
 import { Failure, Result, Success } from "../../core/result";
 import { TYPES } from "../../di/types";
-import type { UserRepository } from "../../domain/repositories/user-repository";
+import type { UserRepository } from "../../domain/interfaces/repositories/user-repository";
 import {
   DeleteUserUseCase,
   DeleteUserUseCaseParam,
-} from "../../domain/usecases/delete-user-usecase";
+} from "../../domain/interfaces/usecases/delete-user-usecase";
 
 @injectable()
 export class DeleteUserUseCaseImpl implements DeleteUserUseCase {
