+++ src/domain/interfaces/usecases/read-user-usecase.ts
@@ -1,5 +1,5 @@
-import { Result } from "../../core/result";
-import { User } from "../entities/user";
+import { Result } from "../../../core/result";
+import { User } from "../../entities/user";
 import { UseCase } from "./usecase";
 
 export interface ReadUserUseCase
