+++ src/domain/interfaces/usecases/usecase.ts
